+++ test/projects/with-composer/src/NewNs/MyClass.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace NewNs;
+
+class MyClass {
+
+}