+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Next
 
+Features:
+
+- Add the command `ede-php-autoload-reload-autoloads` to refresh the
+  composer autoloads when the composer configuration changed
+
+Fixes:
+
 - Always load the root project when visiting a third-party source
-- Improve ecukes tests
+- Improve ecukes tests reliability